+++ packages/annotation/src/index.tsx
@@ -1,5 +1,6 @@
 export * from './arrows'
 export * from './boxes'
+export * from './filters'
 export * from './flowchart'
 export * from './lines'
 export * from './menu'